pyproject.toml
[project]
name = "knowledge-vault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "openai>=1.86.0",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.1",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
